extracted pause menu
